# Introduce StringBuilderPool